Replace checkers.flake8 with checkers.flake8_checker which uses flake8's internals directly.
